+++ examples/pmchwt_single_trace.jl
@@ -106,8 +106,9 @@ end
     Nd = BEAST.nedelec(Σ, Edges)
     RT = [n×BEAST.nedelec(∂Ωᵢ, e) for (∂Ωᵢ,e) in zip(∂Ω, edges)]
 
-    U = BEAST.DirectProductSpace([rt × rt for rt in RT])
-    V = BEAST.DirectProductSpace([Nd × Nd])
+    U = ∏(rt × rt for rt in RT)
+    V = ∏(Nd × Nd)
+    # V = BEAST.DirectProductSpace([Nd × Nd])
 
     return (;U, V)
 end
@@ -228,80 +229,3 @@ PlotlyJS.add_trace!(plt, hm2, row=1, col=2);
 PlotlyJS.add_trace!(plt, hm3, row=2, col=1);
 PlotlyJS.add_trace!(plt, hm4, row=2, col=2);
 display(plt)
-
-error()
-## Debugging
-
-# ds = make(discretization; h=0.08, κ=2.0)
-ds = make(discretization; h=0.15, κ=2.0)
-
-using LinearAlgebra
-using BEAST.LinearMaps
-using BEAST.BlockArrays
-using BEAST.BlockArrays.ArrayLayouts
-using SparseArrays
-
-(;linmaps, vectors) = ds
-(;Auu, RUv) = linmaps
-(;bu) = vectors
-
-MA1 = AbstractMatrix(Auu)
-MR1 = AbstractMatrix(RUv)
-# vb = AbstractArray(bu)
-
-MA2 = Matrix(Auu)
-MR2 = sparse(RUv)
-
-MulAB = ArrayLayouts.Mul(MA1, MR1)
-@which ArrayLayouts.materialize(MulAB)
-InstAB = ArrayLayouts.instantiate(MulAB)
-@which copy(InstAB)
-@which ArrayLayouts.mulreduce(InstAB)
-redAB = ArrayLayouts.mulreduce(InstAB)
-@which copy(redAB)
-simAB = similar(redAB)
-@which copyto!(simAB, redAB)
-
-M = redAB
-dest = simAB
-fct = ArrayLayouts._fill_copyto!(dest, M.C)
-@which muladd!(M.α, M.A, M.B, M.β, fct; Czero = M.Czero)
-error()
-
-@enter copyto!(simAB, redAB)
-
-@time MA2 * MR2
-@time MA1 * MR1
-
-axs1 = axes(Auu)
-axs2 = Tuple( blockedrange([sz[1] for sz in blocksizes(ax1)]) for ax1 in axs1 )
-
-T = ComplexF64
-Y1 = similar(Array{T}, axs1)
-Y2 = similar(Array{T}, axs2)
-
-@which LinearAlgebra.rmul!(Y1, true)
-@which LinearAlgebra.rmul!(Y2, true)
-
-@profview LinearAlgebra.rmul!(Y1, true)
-@profview LinearAlgebra.rmul!(Y2, true)
-
-lmap = Auu.maps[1].lmap
-P = getindex(axes(lmap,1), lmap.I)
-Q = getindex(axes(lmap,2), lmap.J)
-
-Y1IJ = view(Y1, P, Q)
-Y2IJ = view(Y2, P, Q)
-
-@which LinearAlgebra.rmul!(Y1IJ, true)
-@which LinearAlgebra.rmul!(Y2IJ, true)
-
-@profview LinearAlgebra.rmul!(Y1IJ, true)
-@profview LinearAlgebra.rmul!(Y2IJ, true)
-
-# error()
-@time LinearMaps._unsafe_mul!(Y1, Auu, false)
-@profview LinearMaps._unsafe_mul!(Y1, Auu, false)
-# @profview AbstractMatrix{T}(Auu)
-
-error()