+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Cell coloring based lock-free multi-threaded assembly for frequency domain integral operators
 - Higher order Lagrange elements (cx and c0) on segments
 - All LinearMaps can be cast into bilinear forms. Assembly of the latter trivially returns the underlying LinearMap. This enables for example the construction inverses that explicitly exploit block diagonal structure.
+- Use CompScienceMeshes 0.11 providing typed cells.
 
 # New in 2.8.0
 - Improved support of for the Helmholtz equation.